+++ embedded/include/can.h
@@ -12,7 +12,7 @@
 CAN_HandleTypeDef hcan;
 uint8_t message_num;
 
-char* can_read(CAN_HandleTypeDef *hcan);
+int can_read(CAN_HandleTypeDef *hcan);
 int can_send(uint32_t can_id, size_t length, uint8_t* data, CAN_HandleTypeDef *hcan);
 uint8_t * can_send_obd2(uint16_t can_id, size_t length, uint8_t mode,uint16_t pid, CAN_HandleTypeDef *hcan);
 int can_heartbeat( uint8_t *message_number, CAN_HandleTypeDef *hcan);
